Revert "drop_monitor: Require 'CAP_SYS_ADMIN' when joining "events" group"
This reverts commit 855a2b559df16b0cf7b5dcc1a31042347afcabf2 which is
commit e03781879a0d524ce3126678d50a80484a513c4b upstream.

Well, it reverts portions of it, the main portion was already removed in
the merge of 4.19.302 due to other changes happening in the
net/netlink/genetlink.c file.  The main, abi breakage, was removed here
as well.  If all of this is needed in the Android 4.19 tree, it can be
brought back in an abi-safe way by applying the original patch, NOT by
reverting this revert, as that will not work properly.

Bug: 161946584
Change-Id: I4a2384672c0258853ffbc26416b426b3aafd809f
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>